Version 1.1 submitted to app store
Minor bug fixes
Added the capability to view all TapForAll’s business partners, so not only you can look at it in a list, as opposed to a map format, but also see all of them even when they are not in your vicinity.
Optimized the backend system for the features scheduled for the next release

+++ MainSourceCode/ChatMessagesViewController.m
@@ -101,8 +101,9 @@
 #pragma mark -
 #pragma mark UITableViewDataSource
 
-- (int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
+- (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
     return [DataModel sharedDataModelManager].messages.count;
+
 }
 
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath {
